Use Long instead of Int for datasource/destination version
This let's us use the same codec in sdbe that we were using previously

+++ api/src/main/scala/quasar/api/datasource/DatasourceType.scala
@@ -26,7 +26,7 @@ import scalaz.std.tuple._
 import scalaz.syntax.show._
 
 @Lenses
-final case class DatasourceType(name: String, version: Int)
+final case class DatasourceType(name: String, version: Long)
 
 object DatasourceType extends DatasourceTypeInstances
